Move convertToScalaIterator to Utils package
