--- src/main/scala/com/ebiznext/comet/schema/handlers/package.scala
@@ -1,43 +1,0 @@
-/*
- *
- *  * Licensed to the Apache Software Foundation (ASF) under one or more
- *  * contributor license agreements.  See the NOTICE file distributed with
- *  * this work for additional information regarding copyright ownership.
- *  * The ASF licenses this file to You under the Apache License, Version 2.0
- *  * (the "License"); you may not use this file except in compliance with
- *  * the License.  You may obtain a copy of the License at
- *  *
- *  *    http://www.apache.org/licenses/LICENSE-2.0
- *  *
- *  * Unless required by applicable law or agreed to in writing, software
- *  * distributed under the License is distributed on an "AS IS" BASIS,
- *  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  * See the License for the specific language governing permissions and
- *  * limitations under the License.
- *
- *
- */
-
-package com.ebiznext.comet.schema
-
-import org.apache.hadoop.fs.RemoteIterator
-
-import scala.language.implicitConversions
-
-package object handlers {
-
-  /** Make us able to use a java Hadoop Iterator as a scala Iterator
-    * Making us able to use filter, map ...
-    * @param underlying the java Hadoop Iterator
-    * @tparam T
-    * @return the augmented iterator
-    */
-  implicit def convertToScalaIterator[T](underlying: RemoteIterator[T]): Iterator[T] = {
-    case class wrapper(underlying: RemoteIterator[T]) extends Iterator[T] {
-      override def hasNext = underlying.hasNext
-
-      override def next = underlying.next
-    }
-    wrapper(underlying)
-  }
-}
